dispatch and message tests passing
