+++ codec_test.go
@@ -7,7 +7,6 @@ import (
 
 	"github.com/stretchr/testify/require"
 	"github.com/ugorji/go/codec"
-	"golang.org/x/net/context"
 )
 
 // This test determines the behavior of codec with respect to advancing the
@@ -49,47 +48,4 @@ func TestCodec(t *testing.T) {
 	require.Nil(t, err, "expected decoding to succeed")
 	require.Equal(t, math.MaxInt32, targetInt, "expected codec to successfully decode int")
 	require.Equal(t, 0, len(buf.Bytes()), "expected buffer to be empty")
-}
-
-func TestCodecStruct(t *testing.T) {
-	var buf bytes.Buffer
-	mh := &codec.MsgpackHandle{
-		WriteExt: true,
-	}
-	enc := codec.NewEncoder(&buf, mh)
-	dec := codec.NewDecoder(&buf, mh)
-
-	v := []interface{}{MethodCall, 999, "abc.hello", new(interface{})}
-
-	err := enc.Encode(v)
-	require.Nil(t, err, "expected encoding to succeed")
-	require.Equal(t, 16, len(buf.Bytes()), "expected buffer to contain bytes")
-
-	// Advance the buffer past the msgpack fixarray descriptor byte
-	b, _ := buf.ReadByte()
-	nb := int(b)
-	require.Equal(t, 0x94, nb)
-
-	p := newProtocolHandler(nil)
-	p.registerProtocol(Protocol{
-		Name: "abc",
-		Methods: map[string]ServeHandlerDescription{
-			"hello": {
-				MakeArg: func() interface{} {
-					return nil
-				},
-				Handler: func(context.Context, interface{}) (interface{}, error) {
-					return nil, nil
-				},
-				MethodType: MethodCall,
-			},
-		},
-	})
-	c := RPCCall{}
-	err = c.Decode(nb-0x90, dec, p, nil)
-	require.Nil(t, err)
-	require.Equal(t, MethodCall, c.Type())
-	require.Equal(t, seqNumber(999), c.SeqNo())
-	require.Equal(t, "abc.hello", c.Name())
-	require.Equal(t, nil, c.Arg())
 }
