+++ dispatch.go
@@ -1,6 +1,7 @@
 package rpc
 
 import (
+	"fmt"
 	"io"
 
 	"golang.org/x/net/context"
@@ -56,6 +57,7 @@ func (d *dispatch) Call(ctx context.Context, name string, arg interface{}, res i
 			return err
 		}
 	case <-c.ctx.Done():
+		fmt.Printf("returning from call\n")
 		return d.handleCancel(c)
 	case <-d.stopCh:
 		return io.EOF
