+++ dispatch_test.go
@@ -39,26 +39,34 @@ func dispatchTestCall(t *testing.T) (dispatcher, *callContainer, chan error) {
 	return dispatchTestCallWithContext(t, context.Background())
 }
 
+func sendResponse(c *call, err error) {
+	c.resultCh <- &RPCCall{
+		typ: MethodResponse,
+		RPCData: &RPCResponseData{
+			err: err,
+			c:   c,
+		},
+	}
+}
+
 func TestDispatchSuccessfulCall(t *testing.T) {
 	d, calls, done := dispatchTestCall(t)
 
-	c := calls.retrieveCall(0)
+	c := calls.RetrieveCall(0)
 	require.NotNil(t, c, "Expected c not to be nil")
 
-	ok := c.Finish(nil)
-	require.True(t, ok, "Expected c.Finish to succeed")
+	sendResponse(c, nil)
 	err := <-done
 	require.Nil(t, err, "Expected no error")
 
-	closed := d.Close(nil)
-	<-closed
+	d.Close()
 }
 
 func TestDispatchCanceledBeforeResult(t *testing.T) {
 	ctx, cancel := context.WithCancel(context.Background())
 	d, calls, done := dispatchTestCallWithContext(t, ctx)
 
-	c := calls.retrieveCall(0)
+	c := calls.RetrieveCall(0)
 	require.NotNil(t, c, "Expected c not to be nil")
 
 	cancel()
@@ -67,37 +75,32 @@ func TestDispatchCanceledBeforeResult(t *testing.T) {
 	_, canceled := err.(CanceledError)
 	require.True(t, canceled, "Expected rpc.CanceledError")
 
-	ok := c.Finish(nil)
-	require.False(t, ok, "Expected c.Finish to fail")
+	require.Nil(t, calls.RetrieveCall(0), "Expected call to be removed from the container")
 
-	closed := d.Close(nil)
-	<-closed
+	d.Close()
 }
 
 func TestDispatchCanceledAfterResult(t *testing.T) {
 	ctx, cancel := context.WithCancel(context.Background())
 	d, calls, done := dispatchTestCallWithContext(t, ctx)
 
-	c := calls.retrieveCall(0)
+	c := calls.RetrieveCall(0)
 	require.NotNil(t, c, "Expected c not to be nil")
 
-	ok := c.Finish(nil)
-	require.True(t, ok, "Expected c.Finish to succeed")
+	sendResponse(c, nil)
 
 	cancel()
 
 	err := <-done
-	require.Nil(t, err, "Expected no error")
+	require.Nil(t, err, "Expected call to complete prior to cancel")
 
-	closed := d.Close(nil)
-	<-closed
+	d.Close()
 }
 
 func TestDispatchEOF(t *testing.T) {
 	d, _, done := dispatchTestCall(t)
 
-	closed := d.Close(nil)
-	<-closed
+	d.Close()
 	err := <-done
 	require.Equal(t, io.EOF, err, "Expected EOF")
 }
@@ -105,12 +108,11 @@ func TestDispatchEOF(t *testing.T) {
 func TestDispatchCallAfterClose(t *testing.T) {
 	d, calls, done := dispatchTestCall(t)
 
-	c := calls.retrieveCall(0)
-	c.Finish(nil)
+	c := calls.RetrieveCall(0)
+	sendResponse(c, nil)
 
 	err := <-done
-	closed := d.Close(nil)
-	<-closed
+	d.Close()
 
 	done = runInBg(func() error {
 		return d.Call(context.Background(), "whatever", new(interface{}), new(interface{}), nil)
